+++ NEWS.md
@@ -1,7 +1,8 @@
-### httptest 2.1.3 (under development)
-* Add redacting functions (`redact_auth`, `redact_cookies`, `redact_headers`, `redact_HTTP_auth`) that can be specified in `capture_requests` so that sensitive information like tokens and ids can be purged from recorded response files. The default redacting function is `redact_auth`, which wraps them all.
+## httptest 2.2.0
+* Add redacting functions (`redact_auth`, `redact_cookies`, `redact_HTTP_auth`, `redact_headers`, `within_body_text`) that can be specified in `capture_requests` so that sensitive information like tokens and ids can be purged from recorded response files. The default redacting function is `redact_auth`, which wraps several of them. See the [vignette]((https://github.com/nealrichardson/httptest/blob/master/inst/doc/redacting.md)) for more.
+* When loading a JSON mock response, the current "request" object is now included in the response returned, as is the case with real responses.
 * Remove the file size limitation for mock files loaded in `with_mock_API`
-* `skip_if_disconnected` now also wraps `testthat::skip_on_cran`
+* `skip_if_disconnected` now also wraps `testthat::skip_on_cran` so that tests that require a real network connection don't cause a flaky test failure on CRAN
 
 ### httptest 2.1.2
 * Fix for compatibility with upcoming release of [httr](https://github.com/r-lib/httr/) that affected non-GET requests that did not contain any request body.